% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HigherOrderRiskPreferences.R
\name{tpower}
\alias{tpower}
\title{Truncated p-th power function. Helper function for creating the B-Spline basis (Code by Paul Eilers, Package JOPS, http://statweb.lsu.edu/faculty/marx/JOPS_0.1.0.tar.gz)}
\usage{
tpower(x, t, p)
}
\arguments{
\item{x}{Function value.}

\item{t}{Point of truncation.}

\item{p}{degree of the truncated polynomial function.}
}
\value{
Returns a piece-wise defined basis functions for x > t.
}
\description{
Truncated p-th power function. Helper function for creating the B-Spline basis (Code by Paul Eilers, Package JOPS, http://statweb.lsu.edu/faculty/marx/JOPS_0.1.0.tar.gz)
}
\examples{
tpower(1, 2, 3)
}
